契約リストコンポーネントの切り分け
